requirements.txt
streamlit
pandas
numpy
scikit-learn
python-dotenv
transformers
torch
shap
matplotlib
